Restructured simulation samplers; added a sampler for limiting which inquiries are used
