+++ bcipy/simulator/data/sampler/inquiry_sampler.py
@@ -1,116 +1,17 @@
 import logging
 import random
-from abc import ABC, abstractmethod
 from collections import defaultdict
 from pathlib import Path
-from typing import Callable, Dict, List, Tuple
+from typing import Dict, List, Tuple
 
-import numpy as np
 import pandas as pd
 
-from bcipy.simulator.data.data_engine import QueryFilter, RawDataEngine
+from bcipy.simulator.data.data_engine import RawDataEngine
+from bcipy.simulator.data.sampler.base_sampler import Sampler, format_samples
 from bcipy.simulator.data.trial import Trial
 from bcipy.simulator.util.state import SimState
 
 log = logging.getLogger(__name__)
-
-
-def default_reshaper(eeg_responses: List[np.ndarray]) -> np.ndarray:
-    """Default data reshaper.
-
-    Returns
-    -------
-    ndarray with shape (channel_n, trial_n, sample_n)
-    """
-
-    channels_eeg: List[List[np.ndarray]] = [
-        [] for i in range(len(eeg_responses[0]))
-    ]
-
-    for _i, trial_channels_eeg in enumerate(eeg_responses):
-        for c_i, channel_eeg in enumerate(trial_channels_eeg):
-            channels_eeg[c_i].append(channel_eeg)
-
-    return np.array(channels_eeg)
-
-
-class Sampler(ABC):
-    """Represents a strategy for sampling signal model data from a DataEngine
-    comprised of signal data from one or more data collection sessions."""
-
-    def __init__(self, data_engine: RawDataEngine):
-        self.data_engine: RawDataEngine = data_engine
-        self.model_input_reshaper: Callable = default_reshaper
-
-    @abstractmethod
-    def sample(self, state: SimState) -> List[Trial]:
-        """
-        Query the data engine for a list of trials corresponding to each
-        currently displayed symbol.
-
-        Parameters
-        ----------
-            state - specifies the target symbol and current inquiry (displayed symbols).
-
-        Returns
-        -------
-            a list of Trials with an item for each symbol in the current inquiry.
-        """
-        raise NotImplementedError
-
-    def sample_data(self, state: SimState) -> np.ndarray:
-        """
-        Query for trials and reshape for signal model input according to the
-        provided reshaper.
-
-        Return:
-            ndarray of shape (n_channel, n_trial, n_sample)
-        """
-        trials = self.sample(state)
-        return self.reshaped(trials)
-
-    def sample_with_context(self,
-                            state: SimState) -> Tuple[np.ndarray, List[Trial]]:
-        """
-        Returns
-        -------
-            A tuple of the reshaped data (ndarray of shape (n_channel, n_trial, n_sample)) as
-            well as a list of Trial data (metadata and data not reshaped) for context.
-        """
-        trials = self.sample(state)
-        data = self.reshaped(trials)
-        return data, trials
-
-    def reshaped(self, sample_rows: List[Trial]) -> np.ndarray:
-        """Returns the queried trials reshaped into a format that a model can predict."""
-        return self.model_input_reshaper([trial.eeg for trial in sample_rows])
-
-    def set_reshaper(self, reshaper: Callable):
-        """Set the reshaper"""
-        self.model_input_reshaper = reshaper
-
-    def __str__(self):
-        return f"<{self.__class__.__name__}>"
-
-
-class TargetNontargetSampler(Sampler):
-    """Sampler that that queries based on target/non-target label."""
-
-    def sample(self, state: SimState) -> List[Trial]:
-        sample_rows = []
-        for symbol in state.display_alphabet:
-            filters = self.query_filters(
-                symbol, is_target=(symbol == state.target_symbol))
-            filtered_data = self.data_engine.query(filters, samples=1)
-            sample_rows.append(filtered_data[0])
-
-        log.info(f"Samples:\n{format_samples(sample_rows)}")
-        return sample_rows
-
-    def query_filters(self, symbol: str, is_target: bool) -> List[QueryFilter]:
-        """Expression used to query for a single sample."""
-        # QueryFilter('symbol', '==', symbol)
-        return [QueryFilter('target', '==', int(is_target))]
 
 
 class InquirySampler(Sampler):
@@ -181,8 +82,8 @@ class InquirySampler(Sampler):
         inquiry_n = random.choice(source_inquiries[data_source])
 
         # select all trials for the data_source and inquiry
-        inquiry_df = self.data.loc[(self.data['source'] == data_source) &
-                                   (self.data['inquiry_n'] == inquiry_n)]
+        inquiry_df = self.data.loc[(self.data['source'] == data_source)
+                                   & (self.data['inquiry_n'] == inquiry_n)]
         assert len(inquiry_df) == len(
             inquiry_letter_subset), f"Invalid data source {data_source}"
 
@@ -224,8 +125,3 @@ class InquirySampler(Sampler):
         ]
         log.info(f"EEG Samples:\n{format_samples(rows)}")
         return rows
-
-
-def format_samples(sample_rows: List[Trial]) -> str:
-    """Returns a tabular representation of the sample rows."""
-    return '\n'.join([str(row) for row in sample_rows])
